+++ src/config/types.rs
@@ -10,6 +10,7 @@ arg_enum! {
         Tests,
         Doctests,
         Benchmarks,
+        Examples,
     }
 }
 
@@ -37,6 +38,7 @@ impl From<RunType> for CompileMode {
     fn from(run: RunType) -> Self {
         match run {
             RunType::Tests => CompileMode::Test,
+            RunType::Examples => CompileMode::Build,
             RunType::Doctests => CompileMode::Doctest,
             RunType::Benchmarks => CompileMode::Bench,
         }